refactor(heatmaps): extract HeatmapModal, HeatmapImage, and download lib fn
