+++ src/lib/downloadImage.ts
@@ -1,0 +1,10 @@
+const downloadImage = (imageUrl: string, fileName: string) => {
+  const link = document.createElement("a");
+  link.href = imageUrl;
+  link.download = fileName;
+  document.body.appendChild(link);
+  link.click();
+  document.body.removeChild(link);
+};
+
+export default downloadImage;
